Move functions from util.py to core.py and update references
