+++ tests/test_util.py
@@ -5,6 +5,7 @@ Unit tests for the aac.util module.
 from unittest import TestCase
 
 from aac import util
+from aac.spec import core
 
 
 class TestArchUtil(TestCase):
@@ -12,50 +13,11 @@ class TestArchUtil(TestCase):
     Unit test class for aac.util module.
     """
 
-    def test_get_primitive(self):
-        """
-        Unit test for the util.get_primitive method.
-        """
-        expected_results = ["int", "number", "string", "bool", "file", "date", "map"]
-
-        result = util.get_primitives()
-
-        self.assertCountEqual(result, expected_results)
-
-    def test_get_root_names(self):
-        """
-        Unit test for the util.get_root_names method.
-        """
-        expected_results = ["import", "enum", "data", "model", "usecase", "ext"]
-
-        result = util.get_roots()
-
-        self.assertCountEqual(result, expected_results)
-
-    def test_get_aac_spec(self):
-        """
-        Unit test for the util.get_aac_spec method.
-        """
-
-        aac_data, _ = util.get_aac_spec()
-
-        self.assertGreater(len(aac_data.keys()), 0)
-        self.assertGreater(len(aac_data.keys()), 0)
-
-    def test_get_aac_spec_as_yaml(self):
-        """
-        Unit test for the util.get_aac_spec_as_yaml method.
-        """
-
-        aac_yaml = util.get_aac_spec_as_yaml()
-        self.assertGreater(len(aac_yaml), 0)
-        self.assertGreaterEqual(aac_yaml.count("---"), 14)  # there are 14 base types
-
     def test_get_models_by_type(self):
         """
         Unit test for the util.get_models_by_type method.
         """
-        aac_data, aac_enums = util.get_aac_spec()
+        aac_data, aac_enums = core.get_aac_spec()
         num_data = len(aac_data)
         num_enums = len(aac_enums)
         num_models = 0
